+++ SubProject.cmake
@@ -36,11 +36,6 @@
 # - DISABLE_SUBPROJECTS: when set, does not load sub projects. Useful for
 #   example for continuous integration builds
 # - SUBPROJECT_${name}: If set to OFF, the subproject is not added.
-# - INSTALL_PACKAGES: command line cache variable which will "apt-get", "yum" or
-#   "port install" the known system packages. This variable is unset after this
-#   script is parsed by top level projects.
-#   The packages to install are taken from ${PROJECT_NAME}_${type}_DEPENDS
-#   where type is DEB, RPM or PORT depending on the system.
 # - COMMON_SOURCE_DIR: When set, the source code of subprojects will be
 #   downloaded in this path instead of CMAKE_SOURCE_DIR.
 # A sample project can be found at https://github.com/Eyescale/Collage.git
@@ -73,66 +68,6 @@ set(COMMON_SOURCE_DIR "${CMAKE_SOURCE_DIR}" CACHE PATH
 set(__common_source_dir ${COMMON_SOURCE_DIR})
 get_filename_component(__common_source_dir ${__common_source_dir} ABSOLUTE)
 file(MAKE_DIRECTORY ${__common_source_dir})
-
-function(subproject_install_packages name)
-  if(NOT INSTALL_PACKAGES)
-    return()
-  endif()
-
-  string(TOUPPER ${name} NAME)
-  list(APPEND ${NAME}_DEB_DEPENDS pkg-config git cmake git-review doxygen ccache
-    graphviz ${OPTIONAL_DEBS})
-  set(${NAME}_BUILD_DEBS ${NAME}_DEB_DEPENDS)
-  list(APPEND ${NAME}_DEB_DEPENDS ninja-build lcov cppcheck clang
-    clang-format-3.5) # optional deb packages, not added to build spec
-  list(APPEND ${NAME}_PORT_DEPENDS cppcheck)
-
-  if(CMAKE_SYSTEM_NAME MATCHES "Linux" )
-    # Detecting the package manager to use
-    find_program(__pkg_mng apt-get)
-    if(__pkg_mng)
-      set(__pkg_type DEB)
-    else()
-      find_program(__pkg_mng yum)
-      if(__pkg_mng)
-        set(__pkg_type RPM)
-      endif()
-    endif()
-  elseif(APPLE)
-    find_program(__pkg_mng port)
-  endif()
-
-  if(NOT __pkg_mng)
-    message(WARNING "Could not find the package manager tool for installing dependencies in this system")
-    # Removing INSTALL_PACKAGES so the warning is not printed repeatedly.
-    unset(INSTALL_PACKAGES CACHE)
-    return()
-  else()
-    # We don't want __pkg_mng to appear in ccmake.
-    set(__pkg_mng ${__pkg_mng} CACHE INTERNAL "")
-  endif()
-
-  if(CMAKE_SYSTEM_NAME MATCHES "Linux" AND ${NAME}_${__pkg_type}_DEPENDS)
-    list(SORT ${NAME}_${__pkg_type}_DEPENDS)
-    list(REMOVE_DUPLICATES ${NAME}_${__pkg_type}_DEPENDS)
-    message(
-      "Running 'sudo ${__pkg_mng} install ${${NAME}_${__pkg_type}_DEPENDS}'")
-    execute_process(
-      COMMAND sudo ${__pkg_mng} install ${${NAME}_${__pkg_type}_DEPENDS})
-  endif()
-
-  if(${NAME}_PORT_DEPENDS AND APPLE)
-    list(SORT ${NAME}_PORT_DEPENDS)
-    list(REMOVE_DUPLICATES ${NAME}_PORT_DEPENDS)
-    set(${NAME}_PORT_DEPENDS_UNI)
-    foreach(__port ${${NAME}_PORT_DEPENDS})
-      list(APPEND ${NAME}_PORT_DEPENDS_UNI ${__port} +universal)
-    endforeach()
-    message(
-      "Running 'sudo port install ${${NAME}_PORT_DEPENDS} (+universal)'")
-    execute_process(COMMAND sudo port install -p ${${NAME}_PORT_DEPENDS_UNI})
-  endif()
-endfunction()
 
 function(add_subproject name)
   string(TOUPPER ${name} NAME)
@@ -176,8 +111,6 @@ function(add_subproject name)
     set(${name}_DIR "${CMAKE_BINARY_DIR}/${name}" CACHE PATH
       "Location of ${name} project" FORCE)
   endif()
-
-  subproject_install_packages(${name})
 
   # add the source sub directory to our build and set the binary dir
   # to the build tree
@@ -226,8 +159,6 @@ endmacro()
 
 # Interpret .gitsubprojects
 if(EXISTS "${CMAKE_CURRENT_SOURCE_DIR}/.gitsubprojects")
-  subproject_install_packages(${PROJECT_NAME})
-
   # Gather all subprojects
   set(__subprojects_collect 1)
   set(__subprojects) # all appended on each git_subproject invocation
@@ -319,12 +250,6 @@ if(EXISTS "${CMAKE_CURRENT_SOURCE_DIR}/.gitsubprojects")
       set_target_properties(update PROPERTIES EXCLUDE_FROM_DEFAULT_BUILD ON)
     endif()
     add_dependencies(update ${PROJECT_NAME}-update-git-subprojects)
-  endif()
-
-  if(NOT ${PROJECT_NAME}_IS_SUBPROJECT)
-    # If this variable was given in the command line, ensure that the package
-    # installation is only run in this cmake invocation.
-    unset(INSTALL_PACKAGES CACHE)
   endif()
 
 endif()